refactor: inline diff serialize
