+++ packages/ai-native/src/browser/model/serialize.ts
@@ -1,0 +1,4 @@
+export interface ISerializeState<T> {
+  serializeState: () => T;
+  restoreSerializedState(state: T): void;
+}
